removing old bounties listing
